completely reworked chat response system

+++ src/events/americaya.js
@@ -3,8 +3,6 @@ module.exports = {
   async execute(message) {
     if (!message.guild) return;
     if (message.channel.id !== `1221908615334461460`) return;
-    if (!message.content.startsWith("Hallo :D")) {
-      message.delete();
-    }
+    if (message.content !== ("Hallo :D")) {message.delete()}
   },
 };
